Adding String Calculator Scenario 2

+++ src/main/java/com/osherove/kata/StringCalculator.java
@@ -9,18 +9,18 @@ package com.osherove.kata;
  */
 public class StringCalculator {
 
-	public int add(String input) {
-		int result = 0;
+	public long add(String input) {
+		long result = 0;
 		if (input.isEmpty())
 			return 0;
 		else {
 			String[] numbersInString = input.split(",");
-			int[] numberInInt = new int[numbersInString.length];
+			long[] numberInInt = new long[numbersInString.length];
 			int counter = 0;
 			for (String s : numbersInString)
 				numberInInt[counter++] = Integer.parseInt(s);
 			counter = 0;
-			for (int i : numberInInt)
+			for (long i : numberInInt)
 				result += i;
 		}
 		return result;
